Introduce musicdatamodel
The class MusicDataModel
will incorporate several roles.
First, it will act as a database and store the locations of the music
files. Secondly, it will subclass QAbstractTableModel so it can provide
data for the QTableView in the main window. And thirdly it
expands the pattern and makes use of taglib.
For the future it will hopefully provide plug-in support
for MusicBrainz....